+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ui/MessagesActivity.kt
@@ -58,7 +58,6 @@ import androidx.compose.ui.unit.dp
 import io.getstream.chat.android.compose.sample.ChatApp
 import io.getstream.chat.android.compose.sample.R
 import io.getstream.chat.android.compose.state.mediagallerypreview.MediaGalleryPreviewResultType
-import io.getstream.chat.android.compose.state.messages.attachments.StatefulStreamMediaRecorder
 import io.getstream.chat.android.compose.ui.components.composer.MessageInput
 import io.getstream.chat.android.compose.ui.components.messageoptions.MessageOptionItemVisibility
 import io.getstream.chat.android.compose.ui.components.messageoptions.defaultMessageOptionsState
@@ -76,6 +75,7 @@ import io.getstream.chat.android.compose.ui.theme.MessageComposerTheme
 import io.getstream.chat.android.compose.ui.theme.MessageOptionsTheme
 import io.getstream.chat.android.compose.ui.theme.ReactionOptionsTheme
 import io.getstream.chat.android.compose.ui.theme.StreamColors
+import io.getstream.chat.android.compose.ui.theme.StreamShapes
 import io.getstream.chat.android.compose.ui.theme.StreamTypography
 import io.getstream.chat.android.compose.ui.util.rememberMessageListState
 import io.getstream.chat.android.compose.viewmodel.messages.AttachmentsPickerViewModel
@@ -92,14 +92,8 @@ import io.getstream.chat.android.ui.common.state.messages.list.DeletedMessageVis
 import io.getstream.chat.android.ui.common.state.messages.list.SelectedMessageOptionsState
 import io.getstream.chat.android.ui.common.state.messages.list.SelectedMessageReactionsPickerState
 import io.getstream.chat.android.ui.common.state.messages.list.SelectedMessageReactionsState
-import io.getstream.sdk.chat.audio.recording.DefaultStreamMediaRecorder
-import io.getstream.sdk.chat.audio.recording.MediaRecorderState
-import io.getstream.sdk.chat.audio.recording.StreamMediaRecorder
 
 class MessagesActivity : BaseConnectedActivity() {
-
-    private val streamMediaRecorder: StreamMediaRecorder by lazy { DefaultStreamMediaRecorder(applicationContext) }
-    private val statefulStreamMediaRecorder by lazy { StatefulStreamMediaRecorder(streamMediaRecorder) }
 
     private val factory by lazy {
         MessagesViewModelFactory(
@@ -122,21 +116,30 @@ class MessagesActivity : BaseConnectedActivity() {
         super.onCreate(savedInstanceState)
 
         setContent {
-            val colors = if (isSystemInDarkTheme()) StreamColors.defaultDarkColors() else StreamColors.defaultColors()
+            val isInDarkMode = isSystemInDarkTheme()
+            val colors = if (isInDarkMode) StreamColors.defaultDarkColors() else StreamColors.defaultColors()
             val typography = StreamTypography.defaultTypography()
+            val shapes = StreamShapes.defaultShapes()
+            val messageComposerTheme = MessageComposerTheme.defaultTheme(isInDarkMode, typography, shapes, colors)
             ChatTheme(
+                isInDarkMode = isInDarkMode,
                 colors = colors,
+                shapes = shapes,
                 typography = typography,
                 dateFormatter = ChatApp.dateFormatter,
                 autoTranslationEnabled = ChatApp.autoTranslationEnabled,
                 isComposerLinkPreviewEnabled = ChatApp.isComposerLinkPreviewEnabled,
                 allowUIAutomationTest = true,
-                messageComposerTheme = MessageComposerTheme.defaultTheme(typography).let { messageComposerTheme ->
-                    messageComposerTheme.copy(
-                        attachmentCancelIcon = messageComposerTheme.attachmentCancelIcon.copy(
+                messageComposerTheme = messageComposerTheme.let {
+                    it.copy(
+                        attachmentCancelIcon = it.attachmentCancelIcon.copy(
                             painter = painterResource(id = R.drawable.stream_compose_ic_clear),
                             tint = colors.overlayDark,
                             backgroundColor = colors.appBackground,
+                        ),
+                        audioRecording = it.audioRecording.copy(
+                            enabled = true,
+                            showRecordButtonOverSend = false,
                         ),
                     )
                 },
@@ -161,27 +164,11 @@ class MessagesActivity : BaseConnectedActivity() {
                     onTapUserMention = { user ->
                         Log.i("MessagesActivity", "user mention tapped: ${user.id}")
                     },
-                    // TODO
-                    // statefulStreamMediaRecorder = statefulStreamMediaRecorder
                 )
 
                 // MyCustomUi()
             }
         }
-    }
-
-    override fun onPause() {
-        super.onPause()
-        if (statefulStreamMediaRecorder.mediaRecorderState.value == MediaRecorderState.RECORDING) {
-            streamMediaRecorder.stopRecording()
-        } else {
-            streamMediaRecorder.release()
-        }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        streamMediaRecorder.stopRecording()
     }
 
     @Composable
@@ -361,7 +348,6 @@ class MessagesActivity : BaseConnectedActivity() {
                 .fillMaxWidth()
                 .wrapContentHeight(),
             viewModel = composerViewModel,
-            statefulStreamMediaRecorder = statefulStreamMediaRecorder,
             integrations = {},
             input = { inputState ->
                 MessageInput(