+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/StreamAttachmentFactories.kt
@@ -20,6 +20,8 @@ import android.content.Context
 import androidx.activity.compose.ManagedActivityResultLauncher
 import androidx.compose.foundation.Image
 import androidx.compose.ui.layout.ContentScale
+import androidx.core.net.toUri
+import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.compose.state.mediagallerypreview.MediaGalleryPreviewResult
 import io.getstream.chat.android.compose.ui.attachments.content.onFileAttachmentContentItemClick
 import io.getstream.chat.android.compose.ui.attachments.content.onFileUploadContentItemClick
@@ -37,6 +39,7 @@ import io.getstream.chat.android.compose.ui.attachments.factory.UploadAttachment
 import io.getstream.chat.android.compose.ui.attachments.preview.MediaGalleryPreviewContract
 import io.getstream.chat.android.compose.ui.attachments.preview.handler.AttachmentPreviewHandler
 import io.getstream.chat.android.compose.ui.theme.StreamDimens
+import io.getstream.chat.android.compose.viewmodel.messages.AudioPlayerViewModelFactory
 import io.getstream.chat.android.models.Attachment
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.ui.common.images.resizing.StreamCdnImageResizing
@@ -57,6 +60,7 @@ public object StreamAttachmentFactories {
     /**
      * Default attachment factories we provide, which can transform image, file and link attachments.
      *
+     * @param getChatClient - A lambda that provides the ChatClient instance.
      * @param linkDescriptionMaxLines - The limit of how long the link attachment descriptions can be.
      * @param giphyInfoType Used to modify the quality and dimensions of the rendered
      * Giphy attachments.
@@ -79,6 +83,7 @@ public object StreamAttachmentFactories {
      * @return A [List] of various [AttachmentFactory] instances that provide different attachments support.
      */
     public fun defaultFactories(
+        getChatClient: () -> ChatClient = { ChatClient.instance() },
         linkDescriptionMaxLines: Int = DEFAULT_LINK_DESCRIPTION_MAX_LINES,
         giphyInfoType: GiphyInfoType = GiphyInfoType.ORIGINAL,
         giphySizingMode: GiphySizingMode = GiphySizingMode.ADAPTIVE,
@@ -107,7 +112,14 @@ public object StreamAttachmentFactories {
         UploadAttachmentFactory(
             onContentItemClick = onUploadContentItemClick,
         ),
-        AudioRecordAttachmentFactory,
+        AudioRecordAttachmentFactory(
+            viewModelFactory = AudioPlayerViewModelFactory(
+                getAudioPlayer = { getChatClient().audioPlayer },
+                hasRecordingUri = { it.upload != null || it.assetUrl != null },
+                getRecordingUri = { it.upload?.toUri()?.toString() ?: it.assetUrl },
+            ),
+            getCurrentUserId = { getChatClient().getCurrentOrStoredUserId() },
+        ),
         LinkAttachmentFactory(
             linkDescriptionMaxLines = linkDescriptionMaxLines,
             onContentItemClick = onLinkContentItemClick,