+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 
 ### ✅ Added
 - Split poll options on the poll attachment to show a fit nubmer of options and a button to open new view with all options.[#5406](https://github.com/GetStream/stream-chat-android/pull/5406)
+- Add view to see poll results. [#5405](https://github.com/GetStream/stream-chat-android/pull/5405)
 
 ### ⚠️ Changed
 
@@ -78,6 +79,7 @@
 ### ✅ Added
 - Add a default theme for `MediaPreviewActivity` and `MediaGalleryPreviewActivity`. [#5391](https://github.com/GetStream/stream-chat-android/pull/5391)
 - Added typing indicator on `ChannelList`. [#5399](https://github.com/GetStream/stream-chat-android/pull/5399)
+- Added support for Audio Recording feature. [#5404](https://github.com/GetStream/stream-chat-android/pull/5404)
 
 ### ⚠️ Changed